requirements.txt
python-telegram-bot[webhooks]==21.6
python-dotenv==1.0.0
httpx==0.27.0
Pillow==10.4.0
pytz>=2023.3
numpy>=1.24.0